The approach is to include following features in the app.
It is really useful  for communication in the costly wold of phone calls and SMS.  Although, many chatting applications are
already there, but we want to get real time experience of how this much fast communication is done in real time for many users
simultaneously. We want to learn messaging and networking protocols in vey secured environment with very lightweight and
attractive layouts.